+++ src/views/note-creation-form/Toast.tsx
@@ -1,50 +1,14 @@
-import {
-  createContext,
-  PropsWithChildren,
-  ReactElement,
-  useCallback,
-  useContext,
-  useState,
-} from "react";
+import { ReactElement, useContext } from "react";
 
-interface ToastProps {
+import { ToastStateContext } from "./ToastProvider";
+
+export interface ToastProps {
   id: number;
   content: string | ReactElement;
 }
 
-type ToastStateContextProps = ToastProps[];
-type ToastActionContextProps = ({ content }: Omit<ToastProps, "id">) => void;
-
-export const ToastStateContext = createContext<ToastStateContextProps>([]);
-
-export const ToastActionContext = createContext<ToastActionContextProps>(
-  () => {},
-);
-
-export const ToastProvider = ({ children }: PropsWithChildren) => {
-  const [toasts, setToasts] = useState<ToastProps[]>([]);
-
-  const addToast = useCallback(({ content }: Omit<ToastProps, "id">) => {
-    const newId = Math.random() * Number.MAX_SAFE_INTEGER;
-    setToasts((prev) => [...prev, { id: newId, content }]);
-
-    const timeoutId = setTimeout(() => {
-      setToasts((prev) => prev.slice(1));
-      clearTimeout(timeoutId);
-    }, 2000);
-  }, []);
-
-  return (
-    <ToastStateContext.Provider value={toasts}>
-      <ToastActionContext.Provider value={addToast}>
-        {children}
-      </ToastActionContext.Provider>
-    </ToastStateContext.Provider>
-  );
-};
-
 export default function Toast() {
-  const toasts = useContext<ToastStateContextProps>(ToastStateContext);
+  const toasts = useContext<ToastProps[]>(ToastStateContext);
 
   return (
     <div className="flex h-[500px] flex-col-reverse gap-4">
